refactor: move models to entity package
