+++ app/src/main/java/com/thechance/qurio/data/remote/mapper/GameCategoryMapper.kt
@@ -1,7 +1,7 @@
 package com.thechance.qurio.data.remote.mapper
 
 import com.thechance.qurio.data.remote.dto.GameCategoryDto
-import com.thechance.qurio.domain.model.GameCategory
+import com.thechance.qurio.domain.entity.GameCategory
 
 fun GameCategoryDto.toEntity() = GameCategory(
     title = name ?: "",
