+++ app/src/main/java/com/thechance/qurio/di/PresentationModule.kt
@@ -1,6 +1,6 @@
 package com.thechance.qurio.di
 
-import com.thechance.qurio.domain.repository.GameRepository
+import com.thechance.qurio.domain.repository.game.GameRepository
 import com.thechance.qurio.presentation.screen.games_screen.GamesPresenter
 import com.thechance.qurio.presentation.screen.played_games_screen.PlayedGamesPresenter
 import dagger.Module
