+++ app/src/main/java/com/thechance/qurio/domain/entity/GameCategory.kt
@@ -1,4 +1,4 @@
-package com.thechance.qurio.domain.model
+package com.thechance.qurio.domain.entity
 
 data class GameCategory(
     val title: String,
