+++ app/src/main/java/com/thechance/qurio/presentation/screen/games_screen/GamesPresenter.kt
@@ -1,7 +1,7 @@
 package com.thechance.qurio.presentation.screen.games_screen
 
-import com.thechance.qurio.domain.model.GameCategory
-import com.thechance.qurio.domain.repository.GameRepository
+import com.thechance.qurio.domain.entity.GameCategory
+import com.thechance.qurio.domain.repository.game.GameRepository
 import com.thechance.qurio.presentation.base.BasePresenter
 
 class GamesPresenter(
